backend/requirements.txt
Django==4.2.14
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
dj-database-url==2.2.0
djangorestframework-simplejwt==5.3.1
